backend/pyproject.toml
[project]
name = "sds-backend"
version = "0.1.0"
description = "Smart Driving School - Booking & AI Support Backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "pydantic-settings>=2.6.0",
    "pgvector>=0.3.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "google-generativeai>=0.8.0",
    "python-multipart>=0.0.9",
    "pypdf>=4.3.0",
    "python-docx>=1.1.0",
    "greenlet>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
